SESCRIPTION and NAMESPACE were changed based on Rcpp.package.skeleton("mmnst")
FFIOX.h is not needed any more (proper .cpp function in scr folder)

the .data call in FrequencyScreePlot is causing issues now, so I have made the variables k, p, and f global varaibles, but this needs to be fixed.

PoissonRDP is now avoiding the .Internal, but it is a little slower (the problem is with R Check, not with running the code. We can use the older version for our purposes before submitting to CRAN. The difference is two lines and the "fast" lines are commented out and replaced but new lines without the .Internal call)

RDPCrossValidation is now not using the .Call function, but it is slower. Again, we can use the slower version for us, but when it comes to CRAN submission, to pass the R Check, we will need to switch to the slower veresion.

The BinCount.cpp file is a (bit slower) cpp version of the graphics:::C_BinCount function.